requirements.txt
discord.py>=2.4.0,<3.0.0
davey>=0.1.5,<1.0.0
PyNaCl>=1.5.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
